+++ packages/backend/src/services/DashboardService/DashboardService.mock.ts
@@ -40,6 +40,8 @@ export const user: SessionUser = {
     ]),
     isActive: true,
     abilityRules: [],
+    createdAt: new Date(),
+    updatedAt: new Date(),
 };
 
 export const space: SpaceTable['base'] = {