+++ packages/frontend/src/features/metricsCatalog/components/MetricCatalogCategoryFormItem.tsx
@@ -11,7 +11,6 @@ import {
     TextInput,
     Tooltip,
     UnstyledButton,
-    useMantineTheme,
 } from '@mantine/core';
 import { useDisclosure, useHover } from '@mantine/hooks';
 import { IconDots, IconTrash } from '@tabler/icons-react';
@@ -19,8 +18,9 @@ import { useCallback, useState, type FC } from 'react';
 import MantineIcon from '../../../components/common/MantineIcon';
 import { useAppSelector } from '../../sqlRunner/store/hooks';
 import { useDeleteTag, useUpdateTag } from '../hooks/useProjectTags';
-import { getTagColorSwatches } from '../utils/getRandomTagColor';
+import { TAG_COLOR_SWATCHES } from '../utils/getRandomTagColor';
 import { CatalogCategory } from './CatalogCategory';
+import { CatalogCategorySwatch } from './CatalogCategorySwatch';
 
 type EditPopoverProps = {
     hovered: boolean;
@@ -41,7 +41,6 @@ const EditPopover: FC<EditPopoverProps> = ({
     const [opened, { open, close }] = useDisclosure(false);
     const [editName, setEditName] = useState(category.name);
     const [editColor, setEditColor] = useState(category.color);
-    const colors = getTagColorSwatches(useMantineTheme().colors);
 
     const handleClose = useCallback(() => {
         close();
@@ -123,10 +122,10 @@ const EditPopover: FC<EditPopoverProps> = ({
                     />
 
                     <SimpleGrid cols={7} spacing="xs" verticalSpacing="xs">
-                        {colors.map((color) => (
-                            <CatalogCategory
+                        {TAG_COLOR_SWATCHES.map((color) => (
+                            <CatalogCategorySwatch
                                 key={color}
-                                category={{ name: '', color }}
+                                color={color}
                                 onClick={(e) => {
                                     e.stopPropagation();
                                     setEditColor(color);
@@ -198,8 +197,11 @@ export const MetricCatalogCategoryFormItem: FC<Props> = ({
             pos="relative"
             position="apart"
             sx={(theme) => ({
-                borderRadius: theme.radius.sm,
-                '&:hover': { backgroundColor: theme.colors.gray[2] },
+                borderRadius: theme.radius.md,
+                '&:hover': {
+                    backgroundColor: theme.colors.gray[1],
+                    transition: `background-color ${theme.other.transitionDuration}ms ${theme.other.transitionTimingFunction}`,
+                },
             })}
         >
             <UnstyledButton