+++ packages/frontend/src/features/metricsCatalog/components/MetricsCatalogColumns.tsx
@@ -1,8 +1,18 @@
 import { type CatalogField } from '@lightdash/common';
-import { Box, Group, Highlight, HoverCard, Text, Tooltip } from '@mantine/core';
+import {
+    Flex,
+    Group,
+    Highlight,
+    HoverCard,
+    Text,
+    Tooltip,
+} from '@mantine/core';
+import { useHover } from '@mantine/hooks';
+import { IconPlus } from '@tabler/icons-react';
 import MarkdownPreview from '@uiw/react-markdown-preview';
 import { type MRT_ColumnDef } from 'mantine-react-table';
 import { useMemo, type FC, type SVGProps } from 'react';
+import MantineIcon from '../../../components/common/MantineIcon';
 import {
     Description,
     Hash,
@@ -12,6 +22,7 @@ import {
 import { useAppDispatch, useAppSelector } from '../../sqlRunner/store/hooks';
 import { setCategoryPopoverIsClosing } from '../store/metricsCatalogSlice';
 import { CatalogCategory } from './CatalogCategory';
+import { ExploreMetricButton } from './ExploreMetricButton';
 import { MetricChartUsageButton } from './MetricChartUsageButton';
 import { MetricsCatalogCategoryForm } from './MetricsCatalogCategoryForm';
 import { MetricsCatalogColumnName } from './MetricsCatalogColumnName';
@@ -47,25 +58,31 @@ export const MetricsCatalogColumns: MRT_ColumnDef<CatalogField>[] = [
         header: 'Metric',
         enableSorting: true,
         enableEditing: false,
+        size: 400,
         Header: ({ column }) => (
             <HeaderCell Icon={Hash}>{column.columnDef.header}</HeaderCell>
         ),
-        Cell: ({ row, table }) => (
-            <Tooltip
-                label={row.original.tableName}
-                disabled={!row.original.tableName}
-                withinPortal
-                position="right"
-            >
-                <MetricsCatalogColumnName row={row} table={table} />
-            </Tooltip>
-        ),
+        Cell: ({ row, table }) => {
+            return (
+                <Flex justify="space-between" align="center" w="100%">
+                    <Tooltip
+                        label={row.original.tableName}
+                        disabled={!row.original.tableName}
+                        withinPortal
+                        position="right"
+                    >
+                        <MetricsCatalogColumnName row={row} table={table} />
+                    </Tooltip>
+                    <ExploreMetricButton row={row} className="explore-button" />
+                </Flex>
+            );
+        },
     },
     {
         accessorKey: 'description',
         enableSorting: false,
         enableEditing: false,
-        size: 300,
+        size: 500,
         header: 'Description',
         Header: ({ column }) => (
             <HeaderCell Icon={Description}>
@@ -80,11 +97,18 @@ export const MetricsCatalogColumns: MRT_ColumnDef<CatalogField>[] = [
                 disabled={!row.original.description}
             >
                 <HoverCard.Target>
-                    <Text lineClamp={2}>
+                    <Text
+                        lineClamp={2}
+                        c={row.original.description ? 'dark.4' : 'dark.1'}
+                        fz="sm"
+                        fw={400}
+                        lh="150%"
+                    >
                         <Highlight
                             highlight={table.getState().globalFilter || ''}
+                            lh="150%"
                         >
-                            {row.original.description ?? ''}
+                            {row.original.description ?? '-'}
                         </Highlight>
                     </Text>
                 </HoverCard.Target>
@@ -104,7 +128,7 @@ export const MetricsCatalogColumns: MRT_ColumnDef<CatalogField>[] = [
         header: 'Category',
         enableSorting: false,
         enableEditing: true,
-        size: 200,
+        size: 300,
         minSize: 180,
         mantineTableBodyCellProps: () => {
             return {
@@ -132,7 +156,7 @@ export const MetricsCatalogColumns: MRT_ColumnDef<CatalogField>[] = [
             );
 
             return (
-                <Group spacing="two" pos="relative" w="100%" h="100%">
+                <Group spacing="xxs" pos="relative" w="100%" h="100%">
                     {categories.map((category) => (
                         <CatalogCategory
                             key={category.tagUuid}
@@ -163,6 +187,7 @@ export const MetricsCatalogColumns: MRT_ColumnDef<CatalogField>[] = [
             );
         },
         Cell: ({ row, table, cell }) => {
+            const { hovered, ref } = useHover();
             const isCategoryPopoverClosing = useAppSelector(
                 (state) => state.metricsCatalog.popovers.category.isClosing,
             );
@@ -174,9 +199,11 @@ export const MetricsCatalogColumns: MRT_ColumnDef<CatalogField>[] = [
 
             return (
                 // This is a hack to make the whole cell clickable and avoid race conditions with click outside events
-                <Box
+                <Flex
+                    ref={ref}
                     pos="absolute"
-                    p="md"
+                    py={6}
+                    px="md"
                     left={0}
                     top={0}
                     w="100%"
@@ -189,15 +216,28 @@ export const MetricsCatalogColumns: MRT_ColumnDef<CatalogField>[] = [
                         table.setEditingCell(cell);
                     }}
                 >
-                    <Group spacing="two" pos="relative" w="100%" h="100%">
-                        {categories.map((category) => (
-                            <CatalogCategory
-                                key={category.tagUuid}
-                                category={category}
+                    {categories.length === 0 && hovered ? (
+                        <Group spacing={2}>
+                            <MantineIcon
+                                color="dark.1"
+                                icon={IconPlus}
+                                size={12}
                             />
-                        ))}
-                    </Group>
-                </Box>
+                            <Text span fz="sm" color="dark.1">
+                                Click to add
+                            </Text>
+                        </Group>
+                    ) : (
+                        <Group spacing="xxs" pos="relative" w="100%" h="100%">
+                            {categories.map((category) => (
+                                <CatalogCategory
+                                    key={category.tagUuid}
+                                    category={category}
+                                />
+                            ))}
+                        </Group>
+                    )}
+                </Flex>
             );
         },
     },
@@ -206,7 +246,14 @@ export const MetricsCatalogColumns: MRT_ColumnDef<CatalogField>[] = [
         header: 'Popularity',
         enableSorting: true,
         enableEditing: false,
-        size: 100,
+        size: 150,
+        mantineTableBodyCellProps: () => {
+            return {
+                sx: {
+                    justifyContent: 'center',
+                },
+            };
+        },
         Header: ({ column }) => (
             <HeaderCell Icon={Popularity}>{column.columnDef.header}</HeaderCell>
         ),